+++ app/Http/Controllers/ExerciseController.php
@@ -5,17 +5,23 @@ namespace App\Http\Controllers;
 use App\Models\Exercise;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ExerciseController extends Controller
 {
     public function getExercise(\Illuminate\Http\Request $request): JsonResponse {
         $exercise = Exercise::find($request->id);
-        if($exercise) return response()->json(['status' => 'success', 'exercise' => $exercise]);
+        if($exercise) {
+            if($request->user()->cannot('view', $exercise)) {
+                return response()->json(['error' => 'Unauthorized'], 401);
+            }
+            return response()->json(['status' => 'success', 'exercise' => $exercise]);
+        }
         return response()->json(['status' => 'error', 'report' => 'Non è stata trovata nessun attrezzo']);
     }
 
     public function getAllExercises(){
-        return response()->json(['status' => 'success','exercise' => Exercise::all()]);
+        return response()->json(['status' => 'success','exercise' => Exercise::where('user_id', Auth::user()->id)->get()]);
     }
 
     public function createExercise(\Illuminate\Http\Request $request): JsonResponse {
@@ -39,6 +45,9 @@ class ExerciseController extends Controller
         ]);
         $exercise = Exercise::find($request->id);
         if($exercise){
+            if($request->user()->cannot('update', $exercise)) {
+                return response()->json(['error' => 'Unauthorized'], 401);
+            }
             $exercise->name = $request->name ?? $exercise->name;
             $exercise->description = $request->description ?? $exercise->description;
             if($exercise->save()) return response()->json(['status' => 'success']);
@@ -50,6 +59,9 @@ class ExerciseController extends Controller
     public function deleteExercise(\Illuminate\Http\Request $request): JsonResponse {
         $exercise = Exercise::find($request->id);
         if($exercise) {
+            if($request->user()->cannot('delete', $exercise)) {
+                return response()->json(['error' => 'Unauthorized'], 401);
+            }
             if ($exercise->delete()) return response()->json(['status' => 'success']);
         }
         return response()->json(['status' => 'error', 'report' => 'Non è stato possibile cancellare l\'attrezzo']);
